Bump Mapbox GL Native
Bump version.

mapbox-gl-native @ 377a6e42d687c419e6ae1012b8626336f5dfc1b6

+++ include/mbgl/util/constants.hpp
@@ -11,7 +11,7 @@ namespace mbgl {
 
 namespace util {
 
-constexpr float tileSize = 512;
+constexpr double tileSize = 512;
 
 /*
  * The maximum extent of a feature that can be safely stored in the buffer.
