+++ src/mbgl/util/interpolate.cpp
@@ -13,7 +13,8 @@ float interpolationFactor(float base, Range<float> range, float z) {
     } else if (base == 1.0f) {
         return zoomProgress / zoomDiff;
     } else {
-        return (std::pow(base, zoomProgress) - 1) / (std::pow(base, zoomDiff) - 1);
+        return (std::pow(static_cast<double>(base), zoomProgress) - 1) /
+               (std::pow(static_cast<double>(base), zoomDiff) - 1);
     }
 }
